+++ src/messages.rs
@@ -8,5 +8,5 @@ pub fn creat_review_group_init_message(testers: &[String], publisher: &String) -
 }
 
 pub fn appstore_message() -> &'static str {
-    r#"Welcome to the appstore bot!"#
+    r#"Welcome to the xdcstore!"#
 }