+++ tests/test_something.py
@@ -29,7 +29,7 @@ class BotProcess:
 def bot_binary_path():
     for path in [
         Path.cwd() / "target/debug/github-bot",
-        Path.cwd() / "appstore-bot" / "appstore-bot",
+        Path.cwd() / "xdcstore" / "xdcstore",
     ]:
         if path.exists():
             return path
@@ -39,7 +39,7 @@ def bot_binary_path():
 def bot_assets_path():
     for path in [
         Path.cwd() / "bot-data",
-        Path.cwd() / "appstore-bot" / "bot-data",
+        Path.cwd() / "xdcstore " / "bot-data",
     ]:
         if path.exists():
             return path
@@ -94,7 +94,7 @@ def test_welcome_message(acfactory, storebot):
     bot_chat.send_text("hi!")
 
     msg_in = ac1.wait_next_incoming_message()
-    assert msg_in.text == "Welcome to the appstore bot!"
+    assert msg_in.text == "Welcome to the xdcstore!"
 
 
 def test_update(acfactory, storebot):
@@ -107,7 +107,7 @@ def test_update(acfactory, storebot):
 
     msg_in = ac1.wait_next_incoming_message()
     ac1._evtracker.get_matching("DC_EVENT_WEBXDC_STATUS_UPDATE")
-    assert msg_in.text == "Welcome to the appstore bot!"
+    assert msg_in.text == "Welcome to the xdcstore!"
 
     assert msg_in.is_webxdc()
     status_updates = msg_in.get_status_updates()
@@ -136,7 +136,7 @@ def test_import(acfactory, storebot_example):
 
     msg_in = ac1.wait_next_incoming_message()
     ac1._evtracker.get_matching("DC_EVENT_WEBXDC_STATUS_UPDATE")
-    assert msg_in.text == "Welcome to the appstore bot!"
+    assert msg_in.text == "Welcome to the xdcstore!"
 
     assert msg_in.is_webxdc()
     status_updates = msg_in.get_status_updates()